requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.8
bcrypt==4.1.2
gunicorn==21.2.0
